+++ core/src/main/java/cucumber/cli/Main.java
@@ -13,7 +13,7 @@ public class Main {
     }
 
     public static void run(String[] argv, ClassLoader classLoader) throws IOException {
-        RuntimeOptions runtimeOptions = new RuntimeOptions(argv);
+        RuntimeOptions runtimeOptions = new RuntimeOptions(System.getProperties(), argv);
 
         Runtime runtime = new Runtime(new MultiLoader(classLoader), classLoader, runtimeOptions);
         runtime.writeStepdefsJson();